iot-terminal/requirements.txt
python-escpos==3.0
qrcode==7.4.2
Pillow==10.1.0
